refactor(debugger): extract debugger code to `debugger.cpp`
